requirements.txt
azure-functions
gremlinpython==3.2.7